Refactored StreamHealth table
